xarcade2jstick: install udev rule
With the previous change, whenever the device is unplugged the application will exit.
Now it's possible to add a udev rule to launch xarcade2jstick whenever the device is
plugged. This is way simpler than monitor /dev/input for new devices. Obviously this
won't work as well with 2 or more joysticks being unplugged randomly but should work
for the more usual usecase.

Signed-off-by: Aristeu Rozanski <aris@ruivo.org>