Open virtual devices after opening physical device
If the physical device is already in use, such as when xarcade2jstick is already running, we would previously open the virtual devices then fail to open the physical device (already in use) and exit. This caused Emulation Station to crash as a new gamepad appeared and suddenly disappeared.

Now we check that we can open the physical device successfully before opening the virtual devices.