Use systemd service on Jessie
Udev isn't supposed to run long running programs, even if they detach. As of Jessie the program is terminated once event handling ends, so our old approach of running xarcade2jstick no longer works.

This commit adds a systemd service and detects whether systemd is installed (but doesn't detect if it's in use) and chooses whether to use the old way or the systemd way to launch xarcade2jstick.