Change key combos
Revert back to using just P1 keys in a combo, or P2 keys in a combo, but never a combo of P1 and P2. A combo of P1 and P2 is too easy for P1 and P2 to trigger together by accident! Such as both players pressing SELECT at the same time, or START at the same time.

This retains the logic that prevents combo usage from triggering its component keys.